[#7] Refactor: Validator 폴더링 수정
